(PC-14120)[API] chore: rename function which converts utc to venue timezone and move it to new file for sharing with core.educational
